// extensions/vscode/src/daemon/gen/ferretd/workspace/v1/workspace.pb.ts
export interface WorkspaceId {
  value: string;
}

export interface Workspace {
  id?: WorkspaceId | undefined;
  root: string;
}

export interface OpenRequest {
  root: string;
}

export interface OpenResponse {
  workspace?: Workspace | undefined;
}
// ...
export const WorkspaceId: MessageFns<WorkspaceId> = {
  encode(message: WorkspaceId, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.value !== "") {
      writer.uint32(10).string(message.value);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): WorkspaceId {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseWorkspaceId();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.value = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create(base?: DeepPartial<WorkspaceId>): WorkspaceId {
    return WorkspaceId.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<WorkspaceId>): WorkspaceId {
    const message = createBaseWorkspaceId();
    message.value = object.value ?? "";
    return message;
  },
};

function createBaseWorkspace(): Workspace {
  return { id: undefined, root: "" };
}

export const Workspace: MessageFns<Workspace> = {
  encode(message: Workspace, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.id !== undefined) {
      WorkspaceId.encode(message.id, writer.uint32(10).fork()).join();
    }
    if (message.root !== "") {
      writer.uint32(18).string(message.root);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Workspace {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseWorkspace();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.id = WorkspaceId.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.root = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create(base?: DeepPartial<Workspace>): Workspace {
    return Workspace.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<Workspace>): Workspace {
    const message = createBaseWorkspace();
    message.id = (object.id !== undefined && object.id !== null) ? WorkspaceId.fromPartial(object.id) : undefined;
    message.root = object.root ?? "";
    return message;
  },
};
// ...
export const OpenResponse: MessageFns<OpenResponse> = {
  encode(message: OpenResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.workspace !== undefined) {
      Workspace.encode(message.workspace, writer.uint32(10).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): OpenResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseOpenResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.workspace = Workspace.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create(base?: DeepPartial<OpenResponse>): OpenResponse {
    return OpenResponse.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<OpenResponse>): OpenResponse {
    const message = createBaseOpenResponse();
    message.workspace = (object.workspace !== undefined && object.workspace !== null)
      ? Workspace.fromPartial(object.workspace)
      : undefined;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends { $case: string } ? { [K in keyof Omit<T, "$case">]?: DeepPartial<T[K]> } & { $case: T["$case"] }
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}